dynamic import the custom environment
