+++ src/pages/tezos/customizebot.js
@@ -69,9 +69,10 @@ import legs3 from '../../assets/CryptobotImages/Leg/03xlegs.png';
 import legs4 from '../../assets/CryptobotImages/Leg/04xlegs.png';
 import legs5 from '../../assets/CryptobotImages/Leg/05xlegs.png';
 
-//Custom Environment function
-import { UnsignedByteType, PMREMGenerator } from 'three';
-import { RGBELoader } from 'three/examples/jsm/loaders/RGBELoader.js';
+const CustomEnvironment = Loadable({
+  loader: () => import('src/components/CustomEnvironment'),
+  loading: () => Loading,
+});
 
 const state = {
   current: null,
@@ -324,28 +325,6 @@ const CustomAmbientLight = ({ setImage, grabImage }) => {
   return <ambientLight intensity={0.5} />;
 };
 
-const CustomEnvironment = () => {
-  const { gl, scene } = useThree();
-  const pmremGenerator = new PMREMGenerator(gl);
-  const loader = new RGBELoader();
-  loader.setDataType(UnsignedByteType);
-  pmremGenerator.compileEquirectangularShader();
-
-  useEffect(() => {
-    loader.load('/royal_esplanade_1k_compressed_50ppi.hdr', texture => {
-      const envMap = pmremGenerator.fromEquirectangular(texture).texture;
-
-      scene.environment = envMap;
-      // one can also set scene.background to envMap here
-
-      texture.dispose();
-      pmremGenerator.dispose();
-    });
-  }, [scene, loader, pmremGenerator]);
-
-  return null;
-};
-
 //INTRO TO CUSTOMIZER TOUR
 const steps = [
   {
@@ -473,7 +452,6 @@ const Customizer = ({ location }) => {
       floor: '#ffffff',
     },
   });
-
 
   function uploadData() {
     upload3dModel(
@@ -1041,8 +1019,8 @@ const Customizer = ({ location }) => {
                       getMeshName={getMeshName}
                       setBotColors={setBotColors}
                     />
-                    <CustomEnvironment />
                   </Suspense>
+                  <CustomEnvironment />
                   <OrbitControls enableZoom={true} />
                 </Canvas>
                 <Loading containerStyles={{ background: 'rgba(55, 65, 81)' }} />
